flask-postgres-docker/app/requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
gunicorn==22.0.0

